code/loader/requirements.txt
# requirements.txt
kafka-python
SPARQLWrapper
pystardog
rdflib
rdflib-jsonld
flask
pytz
